+++ api/generated-index.js
@@ -4,6 +4,7 @@ import { readFileSync } from "fs"
 import { join, dirname } from "path"
 import { fileURLToPath } from "url"
 import he from "he"
+import { handleUserProfile } from "../src/lib/seo/handleUserProfile.js"
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = dirname(__filename)
@@ -157,39 +158,6 @@ async function handleCustomPackageHtml(req, res) {
   res.status(200).send(html)
 }
 
-async function handleUserProfile(req, res) {
-  const [_, username] = req.url.split("?")[0].split("/")
-
-  if (!username) {
-    throw new Error("Username not provided")
-  }
-
-  try {
-    const description = he.encode(
-      `Discover the circuits created by ${username} on tscircuit`,
-    )
-
-    const title = he.encode(`${username} - tscircuit`)
-
-    const html = getHtmlWithModifiedSeoTags({
-      title,
-      description,
-      canonicalUrl: `https://tscircuit.com/${he.encode(username)}`,
-      imageUrl: `https://github.com/${username}.png`,
-    })
-
-    res.setHeader("Content-Type", "text/html; charset=utf-8")
-    res.setHeader("Cache-Control", cacheControlHeader)
-    res.setHeader("Vary", "Accept-Encoding")
-    res.status(200).send(html)
-  } catch (error) {
-    if (error.response?.status === 404) {
-      throw new Error("GitHub user not found")
-    }
-    throw error
-  }
-}
-
 async function handleCustomPage(req, res) {
   const [_, page] = req.url.split("?")[0].split("/")
 
@@ -218,7 +186,7 @@ export default async function handler(req, res) {
   }
 
   try {
-    await handleUserProfile(req, res)
+    await handleUserProfile(req, res, getHtmlWithModifiedSeoTags)
     return
   } catch (e) {
     console.warn("Not a user profile:", e.message)
